requirements.txt
bs4==0.0.1
folium==0.14.0
numpy==1.23.5
pandas==1.5.3
plotly-express==0.4.1
requests-html==0.10.0
scikit-learn==1.2.1
streamlit==1.23.1
streamlit-folium==0.12.0
